+++ src/app/components/google-signin/google-signin.component.html
@@ -1,1 +1,3 @@
-<p>google-signin works!</p>
+<button class="google-btn" (click)="loginWithGoogle()">
+    <img src="https://upload.wikimedia.org/wikipedia/commons/c/c1/Google_%22G%22_logo.svg" alt="Google" style="display: flex; justify-content: center;"/>
+</button>
