+++ src/app/tec-libras-client/cadastro/cadastro.component.html
@@ -108,10 +108,8 @@
                 </form>
                 <p style="text-align: center;">Ou entre com:</p>
 
-                <!--Botao login com Google-->
-                    <button class="google-btn" (click)="loginWithGoogle()">
-                      <img src="https://upload.wikimedia.org/wikipedia/commons/c/c1/Google_%22G%22_logo.svg" alt="Google" style="display: flex; justify-content: center;"/>
-                    </button>
+                <app-google-signin></app-google-signin>
+
                 <div style=" text-align: center;">
                   <a class="help_entrar" (click)="paglogin()">Já possuo uma conta</a>
                 </div>
