+++ src/app/tec-libras-client/cadastro/cadastro.component.ts
@@ -4,14 +4,14 @@ import { ThemeService } from '../../services/theme.service';
 import { UserService } from '../../services/user.service';
 import { Router } from '@angular/router';
 import { AuthService } from '../../services/auth.service';
-import { generate } from 'generate-password-ts';
+import { GoogleSigninComponent } from '../../components/google-signin/google-signin.component';
 
 
 
 
 @Component({
   selector: 'app-cadastro',
-  imports: [FormsModule],   
+  imports: [FormsModule, GoogleSigninComponent],   
   templateUrl: './cadastro.component.html',
   styleUrls: ['./cadastro.component.scss']
 })
@@ -59,29 +59,7 @@ export class CadastroComponent {
   isGoogleLogin: boolean = false;
   isSending: boolean = false;
 
-  loginWithGoogle(): void {
-    this.authService.loginWithGoogle()
-     
-    .then(() => {
-      const userDataString = localStorage.getItem('userData') || '{}';
-      const userData = JSON.parse(userDataString);
-
-      this.user.full_name = userData.full_name || '';
-      this.user.user_name = userData.user_name ? userData.user_name.split('@')[0] : '';
-      this.user.email = userData.user_name || '';
-      this.user.password = generate({length: 10,numbers: true,});
-
-      this.isGoogleLogin = true;
-      this.cadastrar();
-    })
-    .catch(err => {
-      console.error('Erro no login com Google:', err);
-    });
-  }
-  
-
-    cadastrar(): void {
-
+  cadastrar(): void {
       if (this.isSending) return;
       if (!this.isGoogleLogin) {
         
@@ -117,7 +95,6 @@ export class CadastroComponent {
       this.UserService.createUser(this.user).subscribe({
         next: (response: any) => {
           console.log(response);
-         
         },
 
         /* erros de validação do backend */
