+++ src/app/tec-libras-client/comunidade/comunidade.component.html
@@ -1,1 +1,3 @@
+<app-header></app-header>
 <p>comunidade works!</p>
+<app-footer></app-footer>
